+++ apps/native-component-list/src/screens/Camera/CameraScreen.tsx
@@ -223,7 +223,7 @@ export default class CameraScreen extends React.Component<object, State> {
       this.camera?.current?.stopRecording();
       return Promise.resolve();
     } else {
-      return this.camera?.current?.recordAsync({ codec: 'hvc1' });
+      return this.camera?.current?.recordAsync();
     }
   };
 
@@ -231,10 +231,7 @@ export default class CameraScreen extends React.Component<object, State> {
     try {
       const result = await this.recordVideo();
       this.setState((state) => ({ recording: !state.recording }));
-
       if (result?.uri) {
-        const info = await FileSystem.getInfoAsync(result.uri);
-        console.log({ info });
         await FileSystem.moveAsync({
           from: result.uri,
           to: `${FileSystem.documentDirectory}photos/${Date.now()}.${result.uri.split('.')[1]}`,
@@ -449,13 +446,11 @@ export default class CameraScreen extends React.Component<object, State> {
           animateShutter
           mirror={this.state.mirror}
           pictureSize={this.state.pictureSize}
-          videoBitrate={1_500_000}
           flash={this.state.flash}
           active
           mode={this.state.mode}
           mute={this.state.mute}
           zoom={this.state.zoom}
-          videoQuality="720p"
           onMountError={this.handleMountError}
           barcodeScannerSettings={{
             barcodeTypes: ['qr', 'pdf417'],