+++ packages/expo-camera/android/src/main/java/expo/modules/camera/CameraViewModule.kt
@@ -205,13 +205,13 @@ class CameraViewModule : Module() {
         view.mirror = false
       }
 
-      Prop("videBitrate") { view, bitrate: Int? ->
+      Prop("videoBitrate") { view, bitrate: Int? ->
         bitrate?.let {
           view.videoEncodingBitrate = it
           return@Prop
         }
-        if (view.videoEncodingBitrate != 10_000_000) {
-          view.videoEncodingBitrate = 10_000_000
+        if (view.videoEncodingBitrate != null) {
+          view.videoEncodingBitrate = null
         }
       }