+++ packages/expo-camera/ios/Current/CameraView.swift
@@ -2,8 +2,6 @@ import UIKit
 import ExpoModulesCore
 import CoreMotion
 
-let DEFAULT_VIDEO_BITRATE = 10_000_000
-
 public class CameraView: ExpoView, EXAppLifecycleListener,
   AVCaptureFileOutputRecordingDelegate, AVCapturePhotoCaptureDelegate, EXCameraInterface, CameraEvent {
   public var session = AVCaptureSession()
@@ -57,7 +55,7 @@ public class CameraView: ExpoView, EXAppLifecycleListener,
     }
   }
 
-  var videoBitrate = DEFAULT_VIDEO_BITRATE
+  var videoBitrate: Int?
 
   var presetCamera = AVCaptureDevice.Position.back {
     didSet {
@@ -591,10 +589,11 @@ public class CameraView: ExpoView, EXAppLifecycleListener,
     if let codec = options.codec {
       let codecType = codec.codecType()
       if videoFileOutput.availableVideoCodecTypes.contains(codecType) {
-        videoFileOutput.setOutputSettings(
-          [AVVideoCodecKey: codecType,
-           AVVideoCompressionPropertiesKey: [AVVideoAverageBitRateKey: videoBitrate]
-          ], for: connection)
+        var outputSettings: [String: Any] = [AVVideoCodecKey: codecType]
+        if let videoBitrate {
+          outputSettings[AVVideoCompressionPropertiesKey] = [AVVideoAverageBitRateKey: videoBitrate]
+        }
+        videoFileOutput.setOutputSettings(outputSettings, for: connection)
         self.videoCodecType = codecType
       } else {
         promise.reject(CameraRecordingException(options.codec?.rawValue))